refactor(circle): extraire MemberAvatarStack pour dédupliquer la rangée d'avatars
La rangée d'avatars superposés avec tooltip au survol était dupliquée
à 3 endroits : page publique (version cliquable + version statique) et
page dashboard. ~25 lignes répétées × 3 = risque de drift à chaque
retouche visuelle.

Extraction en composant dédié src/components/circles/member-avatar-stack.tsx.
Les wrappers cliquables (CircleMembersDialog) et non cliquables (div)
restent en place ; seule la partie visuelle commune est factorisée.

Résultat : -86 lignes, +40 lignes au total, une source unique de vérité
pour l'affichage des avatars membres.
